Implement reduce_word_count and simple tests
The design document explicitly references "map" as the data type for the
various hashes containing word counts. I found an excellent built-in Python
data type, Counter, that is specialized for our use case. For now, I am using
it internally in collector.py, but I think it may make sense to change the data
type project-wide in a later commit.